Move header files to include folder and fix paths
