+++ magtag-common/include/magtag-common/accel.h
