+++ magtag-common/include/magtag-common/json-helper.h
