+++ magtag-common/include/magtag-common/magtag_epaper.h
